Fix route functions and routes. Add /users route and some basic functions
