+++ lib/db.js
@@ -15,7 +15,7 @@ pool.connect(() => {
   console.log('Connected to database');
 });
 
-module.exports = { pool };
+module.exports = pool;
 
 
 // let dbParams = {};
