+++ lib/users-queries.js
@@ -3,9 +3,20 @@ const pool = require('./db');
 const getUsers = () => {
   return pool
     .query('SELECT * FROM users;')
-    .then(res => res.rows)
+    .then(res => { 
+      console.log("res.rows inside users route ", res.rows);
+      return res.rows;
+    })
     .catch(err => err.messages)
 };
+
+const getUsersById = (id) => {
+  const values = [ id ];
+  return pool
+    .query('SELECT * FROM users WHERE id= %1;', values)
+    .then(res => res.rows[0])
+    .catch(err => err.messages)
+}
 
 const getUsersByEmail = (email) => {
   const values = [ email ];
@@ -17,5 +28,6 @@ const getUsersByEmail = (email) => {
 
 module.exports = {
   getUsers,
+  getUsersById,
   getUsersByEmail
 };
