+++ server.js
@@ -1,16 +1,28 @@
+// load .env data into process.env
 require('dotenv').config();
 
-const express = require('express');
-const router = express.Router();
-
-const morgan = require('morgan');
+// Web server config
+const ENV        = process.env.ENV || "development";
+const express    = require("express");
 const bodyParser = require("body-parser");
 const sass       = require("node-sass-middleware");
-const database   = require('./lib/db');
+const morgan     = require('morgan');
 
-const port = process.env.PORT || 8080;
-const app = express();
+const app        = express();
+const PORT       = process.env.PORT || 8080;
 
+const database = require('./lib/db');
+const usersRouter = require('./routes/users-route');
+
+// PG database client/connection setup
+// const { Pool } = require('pg');
+// const dbParams = require('./lib/db.js');
+// const db = new Pool(dbParams);
+// db.connect();
+
+// Load the logger first so all (static) HTTP requests are logged to STDOUT
+// 'dev' = Concise output colored by response status for development use.
+//         The :status token will be colored red for server error codes, yellow for client error codes, cyan for redirection codes, and uncolored for all other codes.
 app.use(morgan('dev'));
 app.set("view engine", "ejs");
 app.use(bodyParser.urlencoded({ extended: true }));
@@ -22,14 +34,26 @@ app.use("/styles", sass({
 }));
 app.use(express.static("public"));
 
+// Separated Routes for each Resource
+// Note: Feel free to replace the example routes below with your own
+const usersRoute = require("./routes/users-route");
+const widgetsRoutes = require("./routes/widgets");
 
 
+// Mount all resource routes
+// Note: Feel free to replace the example routes below with your own
+app.use('/users', usersRouter(database));
+app.use("/api/widgets", widgetsRoutes(database));
+// Note: mount other resources here, using the same pattern above
 
 
-app.get('/', (req, res) => {
-  res.send('Hello');
+// Home page
+// Warning: avoid creating more routes in this file!
+// Separate them into separate routes files (see above).
+app.get("/", (req, res) => {
+  res.render("index");
 });
 
-app.listen(port, () => {
-  console.log('app listening on port ', port);
+app.listen(PORT, () => {
+  console.log(`Example app listening on port ${PORT}`);
 });
